requirements.txt
PyJWT>=2.8
cryptography>=42
requests>=2.31
python-dotenv>=1.0
rich>=13
